move execute_directives into WastRunner
