+++ crates/wasmi/tests/spec/run.rs
@@ -1,7 +1,6 @@
-use super::{error::TestError, TestDescriptor};
+use super::TestDescriptor;
 use crate::spec::runner::{RunnerConfig, WastRunner};
-use anyhow::Result;
-use wast::{lexer::Lexer, parser::ParseBuffer, token::Span, QuoteWat, Wast, WastDirective, Wat};
+use wast::{lexer::Lexer, parser::ParseBuffer};
 
 /// Runs the Wasm test spec identified by the given name.
 pub fn run_wasm_spec_test(name: &'static str, file: &'static str, config: RunnerConfig) {
@@ -34,178 +33,13 @@ pub fn run_wasm_spec_test(name: &'static str, file: &'static str, config: Runner
         }
     };
 
-    execute_directives(&test, wast, &mut context).unwrap_or_else(|error| {
-        panic!(
-            "{}: failed to execute `.wast` directive: {}",
-            test.path(),
-            error
-        )
-    });
-}
-
-fn execute_directives(
-    test: &TestDescriptor,
-    wast: Wast,
-    test_context: &mut WastRunner,
-) -> Result<()> {
-    let mut results = Vec::new();
-    for directive in wast.directives {
-        match directive {
-            #[rustfmt::skip]
-            WastDirective::ModuleDefinition(
-                | mut module @ QuoteWat::Wat(wast::Wat::Module(_))
-                | mut module @ QuoteWat::QuoteModule { .. },
-            ) => {
-                let wasm = module.encode().unwrap();
-                let span = module.span();
-                test_context.module_compilation_succeeds(test, span, None, &wasm);
-            }
-            #[rustfmt::skip]
-            WastDirective::Module(
-                | mut module @ QuoteWat::Wat(wast::Wat::Module(_))
-                | mut module @ QuoteWat::QuoteModule { .. },
-            ) => {
-                let wasm = module.encode().unwrap();
-                let span = module.span();
-                let id = module.name();
-                test_context.module_compilation_succeeds(test, span, id, &wasm);
-            }
-            WastDirective::AssertMalformed {
-                span,
-                module: mut module @ QuoteWat::Wat(wast::Wat::Module(_)),
-                message,
-            } => {
-                let id = module.name();
-                let wasm = module.encode().unwrap();
-                test_context.module_compilation_fails(test, span, id, &wasm, message);
-            }
-            WastDirective::AssertMalformed { .. } => {}
-            #[rustfmt::skip]
-            WastDirective::AssertInvalid {
-                span,
-                module:
-                    | mut module @ QuoteWat::Wat(wast::Wat::Module(_))
-                    | mut module @ QuoteWat::QuoteModule { .. },
-                message,
-            } => {
-                let id = module.name();
-                let wasm = module.encode().unwrap();
-                test_context.module_compilation_fails(test, span, id, &wasm, message);
-            }
-            WastDirective::Register { span, name, module } => {
-                let module_name = module.map(|id| id.name());
-                let instance = test_context
-                    .instance_by_name_or_last(module_name)
-                    .unwrap_or_else(|error| {
-                        panic!("{}: failed to load module: {}", test.spanned(span), error)
-                    });
-                test_context.register_instance(name, instance);
-            }
-            WastDirective::Invoke(wast_invoke) => {
-                let span = wast_invoke.span;
-                test_context
-                    .invoke(test, wast_invoke, &mut results)
-                    .unwrap_or_else(|error| {
-                        panic!(
-                            "{}: failed to invoke `.wast` directive: {}",
-                            test.spanned(span),
-                            error
-                        )
-                    });
-            }
-            WastDirective::AssertTrap {
-                span,
-                exec,
-                message,
-            } => match test_context.execute_wast_execute(test, exec, &mut results) {
-                Ok(results) => panic!(
-                    "{}: expected to trap with message '{}' but succeeded with: {:?}",
-                    test.spanned(span),
-                    message,
-                    results
-                ),
-                Err(error) => assert_trap(test, span, error, message),
-            },
-            WastDirective::AssertReturn {
-                span,
-                exec,
-                results: expected,
-            } => {
-                test_context
-                    .execute_wast_execute(test, exec, &mut results)
-                    .unwrap_or_else(|error| {
-                        panic!(
-                            "{}: encountered unexpected failure to execute `AssertReturn`: {}",
-                            test.spanned(span),
-                            error
-                        )
-                    });
-                test_context.assert_results(test, span, &results, &expected);
-            }
-            WastDirective::AssertExhaustion {
-                span,
-                call,
-                message,
-            } => match test_context.invoke(test, call, &mut results) {
-                Ok(results) => {
-                    panic!(
-                            "{}: expected to fail due to resource exhaustion '{}' but succeeded with: {:?}",
-                            test.spanned(span),
-                            message,
-                            results
-                        )
-                }
-                Err(error) => assert_trap(test, span, error, message),
-            },
-            WastDirective::AssertUnlinkable {
-                span,
-                module: Wat::Module(mut module),
-                message,
-            } => {
-                let id = module.id;
-                let wasm = module.encode().unwrap();
-                test_context.module_compilation_fails(test, span, id, &wasm, message);
-            }
-            WastDirective::AssertUnlinkable { .. } => {}
-            WastDirective::AssertException { span, exec } => {
-                if let Ok(results) = test_context.execute_wast_execute(test, exec, &mut results) {
-                    panic!(
-                        "{}: expected to fail due to exception but succeeded with: {:?}",
-                        test.spanned(span),
-                        results
-                    )
-                }
-            }
-            unsupported => panic!("encountered unsupported Wast directive: {unsupported:?}"),
-        }
-    }
-    Ok(())
-}
-
-/// Asserts that the `error` is a trap with the expected `message`.
-///
-/// # Panics
-///
-/// - If the `error` is not a trap.
-/// - If the trap message of the `error` is not as expected.
-fn assert_trap(test: &TestDescriptor, span: Span, error: TestError, message: &str) {
-    match error {
-        TestError::Wasmi(error) => {
-            assert!(
-                error.to_string().contains(message),
-                "{}: the directive trapped as expected but with an unexpected message\n\
-                    expected: {},\n\
-                    encountered: {}",
-                test.spanned(span),
-                message,
-                error,
-            );
-        }
-        unexpected => panic!(
-            "{}: encountered unexpected error: \n\t\
-                found: '{unexpected}'\n\t\
-                expected: trap with message '{message}'",
-            test.spanned(span),
-        ),
-    }
+    context
+        .execute_directives(&test, wast)
+        .unwrap_or_else(|error| {
+            panic!(
+                "{}: failed to execute `.wast` directive: {}",
+                test.path(),
+                error
+            )
+        });
 }
